Add exceptions for official URLS

+++ main.py
@@ -9,6 +9,7 @@ intents = nextcord.Intents.all()
 
 bot = commands.Bot(command_prefix='->', description='Prefix is -> | Removing scam links one message at a time', intents=intents, allowed_mentions=nextcord.AllowedMentions(everyone=False, roles=False), max_messages=100000)
 
+bot.official_urls = ['dis.gd', 'discord.co', 'discord.com', 'discord.design', 'discord.dev', 'discord.gg', 'discord.gift', 'discord.gifts', 'discord.media', 'discord.new', 'discord.store', 'discord.tools', 'discordapp.com', 'discordapp.net', 'discordmerch.com', 'discordpartygames.com', 'discord-activities.com', 'discordactivities.com', 'discordsays.com', 'discordstatus.com', 'airhorn.solutions', 'airhornbot.com', 'bigbeans.solutions', 'watchanimeattheoffice.com', 'discordapp.io', 'discordcdn.com']
 
 def get_member(GuildID: int, MemberID: int):
     guild = bot.get_guild(GuildID)
@@ -38,7 +39,7 @@ async def on_message(message):
 
     scam_links_1 = content.split("\n")
     for y in scam_links_1:
-        if y.casefold() in message2 and y != '':
+        if y.casefold() in message2 and y != '' and y not in bot.official_urls:
             try:
                 await message.delete()
                 print("Link found in: discord-scam-links")
